add pool tvl query
